requirements-streamlit.txt
# Streamlit app dependencies (no scikit-learn/joblib — ML handled by Model Serving API)
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.6
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
fastf1==3.8.2
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-cloud-storage==2.19.0
google-genai==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
influxdb-client==1.48.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
msgpack==1.1.2
narwhals==2.18.1
numpy==2.2.6
packaging==26.0
pandas==2.3.3
patsy==1.0.2
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
RapidFuzz==3.14.3
referencing==0.37.0
requests==2.33.1
requests-cache==1.3.1
rpds-py==0.30.0
signalrcore==1.0.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
statsmodels==0.14.6
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
timple==0.1.8
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.3
watchdog==6.0.0
websockets==16.0
